% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{apiStatusHERE}
\alias{apiStatusHERE}
\title{Checks HERE REST API status}
\usage{
apiStatusHERE(api_call_object)
}
\arguments{
\item{api_call_object}{The API status object as returned from the httr
package.}
}
\value{
The API status in TRUE or FALSE
}
\description{
Checks HERE REST API status
}
\examples{
\dontrun{
apiStatusHERE()
}
}
