% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBatchGeocodes.R
\name{createBatchJob}
\alias{createBatchJob}
\title{Batch geocode request}
\usage{
createBatchJob(url = setURL(type = "batch geocode"), app_id, app_code,
  indelim = "|", outdelim = "|", body, action = "run", mailto,
  outcols, outputcombined = FALSE)
}
\arguments{
\item{url}{URL to call the API. If no URL is specified the standarc URL
will be used.}

\item{app_id}{App id for the HERE REST API call. See the vignette: Getting
Started how to get an App Id.}

\item{app_code}{App Code for the HERE REST API call. See the vignette: Getting
Started how to get an App Code.}

\item{indelim}{Field delimeter in the input data.}

\item{outdelim}{Field delimeter in the output data.}

\item{body}{Txt file with the data for which the geocodes need te be added.}

\item{action}{Type of request, run is the standard,}

\item{mailto}{Email address where completion notification is sent.}

\item{outcols}{List of columns to return in the output.}

\item{outputcombined}{If true, the output of successful and unsuccessful
(empty response) Geocoder queries are combined into a single
result_YYYYMMDD-HH-MM__out.txt file. If false, they are returned in
two separate files ; result_YYYYMMDD-HH-MM__out.txt and
result_YYYYMMDD-HH-MM__err.txt.}
}
\value{
HERE RequestID
}
\description{
Batch geocode request
}
\examples{
\dontrun{
createBatchJob(url = "xxx", app_id = "xxx", app_code = "xxx", indelim = "|", outdelim = "|", body = "xxx", action = "run", mailto = "test@test.com", outcols = "xxx", outputcombined = FALSE)
}
}
