% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createWaypoint}
\alias{createWaypoint}
\title{Helper function to create a waypoint from a lattitude and a
longitude.}
\usage{
createWaypoint(lat, lon)
}
\arguments{
\item{lat}{Lattitude coordinates}

\item{lon}{Longitude coordinates}
}
\value{
A waypoint geocoordinates object as use is the HERE
routing API>
}
\description{
Helper function to create a waypoint from a lattitude and a
longitude.
}
\examples{
createWaypoint(48.20849, 16.37208)
}
