% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoutes.R
\name{diffDistTruckCar}
\alias{diffDistTruckCar}
\title{Return delta in distance between truck and car routes}
\usage{
diffDistTruckCar(url = setRouteURL(), app_id, app_code, waypoint0,
  waypoint1, unit = "km", rnd = TRUE)
}
\arguments{
\item{url}{The base url as required to connect to the HERE API. If no value
is passed in this parameter the standard parameter from the setRouteURL() is
used. See the documentation for the setRouteURL() for more information about
the base url.}

\item{app_id}{App id for the HERE REST API call. See the vignette: Getting
Started how to get an App Id.}

\item{app_code}{App Code for the HERE REST API call. See the vignette: Getting
Started how to get an App Code.}

\item{waypoint0}{Starting point as a position with in geocoordinates.}

\item{waypoint1}{End point as a position with in geocoordinates.}

\item{unit}{The unit of measure of the outputted distance object. Currently
only "km" and "mi" are an option.}

\item{rnd}{Rounds the result. TRUE is the standard setting.}
}
\value{
A list with the truck and car distance
}
\description{
Return delta in distance between truck and car routes
}
\examples{
\dontrun{
diffDistTruckCar(
app_id = "xxx",
app_code = "xxx",
waypoint0 = "52.5,13.4",
waypoint1 = "54.5,13")
}
}
