% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extractDist}
\alias{extractDist}
\title{Extracts the distance from a getRoute() result.}
\usage{
extractDist(route_object, unit = "km", rnd = TRUE)
}
\arguments{
\item{route_object}{The object as otputted by the getRoute() function.}

\item{unit}{The unit of measure of the outputted distance object. Currently
only "km" and "mi" are an option.}

\item{rnd}{Rounds the result. TRUE is the standard setting.}
}
\value{
The distance from a getRoute() object in km or mi.
}
\description{
Extracts the distance from a getRoute() result.
}
\examples{
route_obj <- list(response.route.summary.distance = 100.25)
extractDist(route_object = route_obj)
}
