% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoutes.R
\name{getRoute}
\alias{getRoute}
\title{Get route information}
\usage{
getRoute(url = setURL(type = "routing"), app_id, app_code, waypoint0,
  waypoint1, vehicle = "truck")
}
\arguments{
\item{url}{The base url as required to connect to the HERE API. If no value
is passed in this parameter the standard parameter from the setRouteURL() is
used. See the documentation for the setRouteURL() for more information about
the base url.}

\item{app_id}{App id for the HERE REST API call. See the vignette: Getting
Started how to get an App Id.}

\item{app_code}{App Code for the HERE REST API call. See the vignette: Getting
Started how to get an App Code.}

\item{waypoint0}{Starting point as a position with in geocoordinates.}

\item{waypoint1}{End point as a position with in geocoordinates.}

\item{vehicle}{Vehicle type for which to calculate the route. Only two
options are possible; "truck" or "car". The paramters defaults to "truck".}
}
\value{
A data.frame with all the information about the requested route.
}
\description{
Get route information
}
\examples{
\dontrun{
getRoute(app_id = "xxx", app_code = "xxx", waypoint0 = "52.5,13.45", waypoint1 = "53.5,13.45")
}
}
