% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setKeys.R
\name{setCode}
\alias{setCode}
\title{Set HERE App ID}
\usage{
setCode(code)
}
\arguments{
\item{code}{The project HERE App Code.}
}
\value{
The checked App Code.
}
\description{
Set HERE App ID
}
\examples{
setCode(code = "xxxxxxxxxxxxxxxxxxxxxx")
}
