% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setKeys.R
\name{setID}
\alias{setID}
\title{Set HERE App ID}
\usage{
setID(id)
}
\arguments{
\item{id}{The project HERE App ID.}
}
\value{
The checked App ID.
}
\description{
Set HERE App ID
}
\examples{
setID(id = "xxxxxxxxxxxxxxxxxxxx")
}
