% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoutes.R
\name{setRouteURL}
\alias{setRouteURL}
\title{Set routing url}
\usage{
setRouteURL(url = "https://route.api.here.com/routing/",
  version = "7.2")
}
\arguments{
\item{url}{The base url for here API. Standard is
"https://route.api.here.com/routing/"}

\item{version}{The version of the API to be called.
Standard is version 7.2}
}
\value{
The desired url for the API call.
}
\description{
Set routing url
}
\examples{
setRouteURL()
}
