% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setURL}
\alias{setURL}
\title{Helper function to set the url in the API calls to the production
standards.}
\usage{
setURL(type)
}
\arguments{
\item{type}{The type of the API call. Can only be "routing" or "geocoding".}
}
\value{
The standard production url for the selected type.
}
\description{
Helper function to set the url in the API calls to the production
standards.
}
\examples{
setURL(type = "routing")
setURL(type = "geocoding")
}
